+++ drift/lib/src/runtime/query_builder/components/order_by.dart
@@ -37,10 +37,6 @@ class OrderingTerm extends Component {
 
   /// How to order NULLs.
   /// When [nulls] is [null], then it's ignored.
-  ///
-  /// Note that this feature are only available in sqlite3 version `3.30.0` and
-  /// newer. When using `sqlite3_flutter_libs` or a web database, this is not
-  /// a problem.
   final NullsOrder? nulls;
 
   /// Creates an ordering term by the [expression], the [mode] (defaults to