+++ drift_flutter/lib/src/native.dart
@@ -10,7 +10,6 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 import 'package:sqlite3/sqlite3.dart';
-import 'package:sqlite3_flutter_libs/sqlite3_flutter_libs.dart';
 
 import 'connect.dart';
 
@@ -55,11 +54,6 @@ QueryExecutor driftDatabase({
 
   return DatabaseConnection.delayed(Future(() async {
     if (!hasConfiguredSqlite) {
-      // Also work around limitations on old Android versions
-      if (Platform.isAndroid) {
-        await applyWorkaroundToOpenSqlite3OnOldAndroidVersions();
-      }
-
       // Make sqlite3 pick a more suitable location for temporary files - the
       // one from the system may be inaccessible due to sandboxing.
       final cachebase = await (native?.tempDirectoryPath?.call() ??